+++ CHANGELOG.md
@@ -2,6 +2,10 @@ NEW FEATURES:
 
 - Handle `metabase_permissions_graph` `view_data` as either a simple string or a serialized JSON object.
 
+BUG FIXES:
+
+- Ignore `aggregation-idents` and `breakout-idents` in the response when they are not part of the card model.
+
 ## 0.8.1 (2024-09-22)
 
 BUG FIXES: